+++ messages/de/motion.php
@@ -28,6 +28,8 @@ return [
     'prevnext_links_next'                => 'Nächster %TYPE%',
     'prevnext_links_prev_res'            => 'Voriger Beschluss',
     'prevnext_links_next_res'            => 'Nächster Beschluss',
+    'prevnext_links_prev_generic'        => 'Voriger',
+    'prevnext_links_next_generic'        => 'Nächster',
     'pdf_version'                        => 'PDF-Version',
     'table_caption'                      => 'Rahmendaten zum Antrag',
     'table_description'                  => 'Diese Tabelle beschreibt den Status, die Antragstellerin und verschiedene Rahmendaten zum Antrag',